requirements.txt
opensearch-py>=2.4.0
openai>=1.0.0
requests>=2.31.0
apscheduler>=3.10.4
pyyaml>=6.0.1
numpy>=1.26.0
sentence-transformers>=2.6.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
rich>=13.7.0
click>=8.1.7
